feat: Allow pulling translations from a commit hash (#62)
* feat: Allow pulling translations from a commit hash

By switching from clone to an explicit init+fetch we can get better control
of what commits are fetched. clone only allows targeting a branch or a
tag, but fetch can ask for any commit. The caveat is that the server needs
to allow this via `uploadpack.allowReachableSHA1InWant` and similar.

Minor changes:

- checkout is now also quiet by default

See https://github.com/openedx/openedx-atlas/issues/48

* chore: run make gengetoptions

* feat: print atlas pull total disk usage for testing purposes

* feat: add tests for bandwidth and disk usage

* chore: run make gengetoptions with v3.3.2

* feat: disable bandwidth calculations on Mac

just in case if it's not compatible

* fix: use actual size, without rouning up to the closest 512

* fix: upgrade github actions to macos-13 to avoid stalling

also updated ubuntu and other actions

---------

Co-authored-by: Tim McCormack <tmccormack@edx.org>